Extend BaseBean with IndexedBaseBean
