+++ Kitodo-DataManagement/src/main/java/org/kitodo/data/database/beans/Project.java
@@ -31,6 +31,7 @@ import javax.xml.bind.annotation.XmlAttribute;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlType;
 
+import org.kitodo.data.database.beans.base.IndexedBaseBean;
 import org.kitodo.data.database.helper.enums.MetadataFormat;
 
 @XmlAccessorType(XmlAccessType.NONE)
@@ -50,7 +51,7 @@ import org.kitodo.data.database.helper.enums.MetadataFormat;
 // "field" as one might expect.
 @Entity
 @Table(name = "project")
-public class Project extends BaseBean implements Comparable<Project> {
+public class Project extends IndexedBaseBean implements Comparable<Project> {
     private static final long serialVersionUID = -8543713331407761617L;
 
     /**
