+++ Kitodo-DataManagement/src/main/java/org/kitodo/data/database/beans/ProjectFileGroup.java
@@ -11,27 +11,19 @@
 
 package org.kitodo.data.database.beans;
 
-import java.io.Serializable;
+import org.kitodo.data.database.beans.base.BaseBean;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.ForeignKey;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
 @Entity
 @Table(name = "projectFileGroup")
-public class ProjectFileGroup implements Serializable {
+public class ProjectFileGroup extends BaseBean {
     private static final long serialVersionUID = -5506252462891480484L;
-
-    @Id
-    @Column(name = "id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
 
     @Column(name = "name")
     private String name;
@@ -54,14 +46,6 @@ public class ProjectFileGroup implements Serializable {
     @ManyToOne
     @JoinColumn(name = "project_id", foreignKey = @ForeignKey(name = "FK_projectFileGroup_project_id"))
     private Project project;
-
-    public Integer getId() {
-        return this.id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
 
     public String getName() {
         return this.name;
