+++ Kitodo-DataManagement/src/main/java/org/kitodo/data/database/beans/Property.java
@@ -22,11 +22,12 @@ import javax.persistence.ManyToMany;
 import javax.persistence.Table;
 import javax.persistence.Transient;
 
+import org.kitodo.data.database.beans.base.IndexedBaseBean;
 import org.kitodo.data.database.helper.enums.PropertyType;
 
 @Entity
 @Table(name = "property")
-public class Property extends BaseBean implements Comparable<Property> {
+public class Property extends IndexedBaseBean implements Comparable<Property> {
     private static final long serialVersionUID = -2356566712752716107L;
 
     @Column(name = "title")
