+++ Kitodo-DataManagement/src/main/java/org/kitodo/data/database/beans/Workpiece.java
@@ -11,6 +11,8 @@
 
 package org.kitodo.data.database.beans;
 
+import org.kitodo.data.database.beans.base.IndexedBaseBean;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -26,7 +28,7 @@ import javax.persistence.Transient;
 
 @Entity
 @Table(name = "workpiece")
-public class Workpiece extends BaseBean {
+public class Workpiece extends IndexedBaseBean {
     private static final long serialVersionUID = 123266825187246791L;
 
     @ManyToOne
