+++ Kitodo-DataManagement/src/main/java/org/kitodo/data/elasticsearch/index/type/BaseType.java
@@ -20,11 +20,7 @@ import java.util.List;
 import org.apache.http.HttpEntity;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
-import org.kitodo.data.database.beans.BaseBean;
-import org.kitodo.data.database.beans.Process;
-import org.kitodo.data.database.beans.Property;
-import org.kitodo.data.database.beans.User;
-import org.kitodo.data.database.beans.UserGroup;
+import org.kitodo.data.database.beans.base.BaseBean;
 import org.kitodo.data.elasticsearch.api.TypeInterface;
 
 /**
